// HELPS/ClientApp/src/types/store/actions/WorkshopActionTypes.ts
export enum WorkshopActionType {
    REQUEST_WORKSHOPS = 'WORKSHOP_ACTION_REQUEST_WORKSHOPS',
    RECEIVE_WORKSHOPS = 'WORKSHOP_ACTION_RECEIVE_WORKSHOPS',
    REQUEST_USER_WORKSHOPS = 'WORKSHOP_ACTION_REQUEST_USER_WORKSHOPS',
    RECEIVE_USER_WORKSHOPS = 'WORKSHOP_ACTION_RECEIVE_USER_WORKSHOPS',
    BOOK_WORKSHOP = 'WORKSHOP_ACTION_BOOK_WORKSHOP',
    RECEIVE_BOOK_WORKSHOP = 'WORKSHOP_ACTION_RECEIVE_BOOK_WORKSHOP',
    WORKSHOP_ERROR = 'WORKSHOP_ACTION_WORKSHOP_ERROR'
}

export interface WorkshopAction {
    type: WorkshopActionType,
    payload?: any
}